Refactor command and reporting
