+++ lib/ruby-on-speed/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module RubyOnSpeed
-  VERSION = '0.9.0'
+  VERSION = '0.10.0'
 end
